Moved recorded reach check file to test folder so it doesn't ship in production. Renamed the EntityRecordingReachTests.java
